+++ packages/orq-rc/RELEASES.md
@@ -1259,3 +1259,13 @@ Based on:
 - [typescript v3.3.0-rc.28] packages/orq-rc
 ### Releases
 - [NPM v3.3.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.28 - packages/orq-rc
+
+## 2025-04-03 06:55:36
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.528.1 (2.565.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.3.0-rc.29] packages/orq-rc
+### Releases
+- [NPM v3.3.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/3.3.0-rc.29 - packages/orq-rc